requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
email-validator==2.0.0
python-dotenv==1.0.0
Werkzeug==3.0.6
SQLAlchemy==2.0.20
gunicorn==23.0.0
psycopg2-binary==2.9.7
pytest==7.4.0
flask
flask-login
flask-sqlalchemy
pandas
scikit-learn
better-profanity
Textblob
pytest
Pillow